move datasetype from dataset
